% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUIFunctions.R
\name{fns}
\alias{fns}
\title{Five Number Summary}
\usage{
fns(vals)
}
\arguments{
\item{vals}{A numeric vector}
}
\value{
A list of the five number summary, outliers, and warnings
}
\description{
Five Number Summary
}
