% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyExtension-package.R
\docType{package}
\name{shinyExtension-package}
\alias{shinyExtension}
\alias{shinyExtension-package}
\title{shinyExtension: Additional UI Functions for Shiny Apps}
\description{
Additional UI functions for shiny that build on existing functions, but with increased versistility and customization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/blukomski99/shinyExtension}
  \item Report bugs at \url{https://github.com/blukomski99/shinyExtension/issues}
}

}
\keyword{internal}
